Final changes to the project structure to include a DTO layer

+++ food-ordering-system/src/main/java/com/jumpstart/foodorderingsystem/repository/CategoryRepository.java
@@ -3,6 +3,10 @@ package com.jumpstart.foodorderingsystem.repository;
 import com.jumpstart.foodorderingsystem.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+/**
+ * Repository responsible for database operations
+ * related to Category entities.
+ */
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
 }
